Move "util" metrics to self-contained endpoints
